+++ collector-amazon-ivs/src/androidTest/java/com/bitmovin/analytics/amazon/ivs/apiv2/PhoneBasicScenariosTest.kt
@@ -5,9 +5,9 @@ import android.os.Looper
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
 import com.amazonaws.ivs.player.Player
-import com.bitmovin.analytics.amazon.ivs.IAmazonIvsPlayerCollector
 import com.bitmovin.analytics.amazon.ivs.IvsPlayerConstants
 import com.bitmovin.analytics.amazon.ivs.IvsTestUtils
+import com.bitmovin.analytics.amazon.ivs.api.IAmazonIvsPlayerCollector
 import com.bitmovin.analytics.example.shared.Samples
 import com.bitmovin.analytics.systemtest.utils.DataVerifier
 import com.bitmovin.analytics.systemtest.utils.EventDataUtils