+++ collector-bitmovin-player/src/test/java/com/bitmovin/analytics/bitmovin/player/IBitmovinPlayerCollectorKotlinTest.kt
@@ -2,6 +2,7 @@ package com.bitmovin.analytics.bitmovin.player
 
 import com.bitmovin.analytics.BitmovinAnalyticsConfig
 import com.bitmovin.analytics.api.AnalyticsConfig
+import com.bitmovin.analytics.bitmovin.player.api.IBitmovinPlayerCollector
 import io.mockk.mockk
 import org.junit.Assert
 import org.junit.Test