+++ collector-exoplayer/src/main/java/com/bitmovin/analytics/exoplayer/api/IExoPlayerCollector.kt
@@ -1,10 +1,11 @@
-package com.bitmovin.analytics.exoplayer
+package com.bitmovin.analytics.exoplayer.api
 
 import android.content.Context
 import com.bitmovin.analytics.BitmovinAnalyticsConfig
 import com.bitmovin.analytics.api.AnalyticsCollector
 import com.bitmovin.analytics.api.AnalyticsConfig
 import com.bitmovin.analytics.api.DefaultMetadata
+import com.bitmovin.analytics.exoplayer.ExoPlayerCollector
 import com.bitmovin.analytics.utils.ApiV3Utils
 import com.google.android.exoplayer2.ExoPlayer
 
@@ -21,7 +22,7 @@ interface IExoPlayerCollector : AnalyticsCollector<ExoPlayer> {
             "Use IExoPlayerCollector.Factory.create(context, analyticsConfig) instead",
             ReplaceWith(
                 "IExoPlayerCollector.Factory.create(context, analyticsConfig)",
-                "com.bitmovin.analytics.exoplayer.IExoPlayerCollector",
+                "com.bitmovin.analytics.exoplayer.api.IExoPlayerCollector",
             ),
         )
         @JvmStatic