+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Upgraded to Kotlin 1.8.20
 - ExoPlayer: More specific PlaybackException mapping (using errorCodes instead of error types and errorCode names are now part of description)
-- Initial v3 API changes
+- Initial v3 API changes, move Collector interfaces into API package
 - Upgrade example app to latest bitmovin player (v3.39.1), latest ivs player (v1.19.0) and latest exoplayer (v2.19.0)
 
 ### Fixed